+++ ListableUI/Tests/AppearanceTests.swift
@@ -16,7 +16,13 @@ class AppearanceTests: XCTestCase
     {
         let appearance = Appearance()
         
-        XCTAssertEqual(appearance.backgroundColor, .white)
+        if #available(iOS 13.0, *) {
+            XCTAssertEqual(appearance.backgroundColor.resolvedColor(with: .init(userInterfaceStyle: .dark)), .black)
+            XCTAssertEqual(appearance.backgroundColor.resolvedColor(with: .init(userInterfaceStyle: .light)), .white)
+        } else {
+            XCTAssertEqual(appearance.backgroundColor, .white)
+        }
+        
         XCTAssertEqual(appearance.showsScrollIndicators, true)
     }
 }